+++ bootstrap-icons-gem/CHANGELOG.md
@@ -14,3 +14,9 @@
 
 - Expose fill and viewBox.
 - Internal refactorings.
+
+## 1.0.0
+
+- Change custom width/height icon behavior.
+  - Before, when custom height provided, the result was `5em` for example, now it will be only `5`, so it will be possible to set the icon size as `px` and not only `em`.
+